Added bulk CSV import, some UI tweaks
