+++ app/Repositories/Transactions.php
@@ -5,16 +5,18 @@ namespace App\Repositories;
 use App\Models\Transaction;
 use App\Models\User;
 use Illuminate\Support\Carbon;
+use Illuminate\Support\Facades\Auth;
 
 class Transactions
 {
-    public function getRecentGrouped(User $user, $limit = 100, $offset = 0): array
+    public function getRecentGrouped(): array
     {
+        $user = Auth::user();
+
         $transactions = Transaction::query()
             ->where('user_id', $user->id)
             ->orderBy('occurred_at', 'desc')
-            ->limit($limit)
-            ->offset($offset)
+            ->limit(10)
             ->get();
 
         $grouped = [];
@@ -23,20 +25,31 @@ class Transactions
             $date = Carbon::createFromTimeString($transaction->occurred_at)
                 ->format('D, d M');
 
-            if ($date === Carbon::today()->format('D, d M')) {
-                $date = 'Today';
-            } elseif ($date === Carbon::yesterday()->format('D, d M')) {
-                $date = 'Yesterday';
+            $year = Carbon::createFromTimeString($transaction->occurred_at)
+                ->format('Y');
+
+            if ($year === Carbon::today()->format('Y')) {
+                if ($date === Carbon::today()->format('D, d M')) {
+                    $group = 'Today';
+                } elseif ($date === Carbon::yesterday()->format('D, d M')) {
+                    $group = 'Yesterday';
+                } else {
+                    $group = $date;
+                }
+            } else {
+                $group = $year;
             }
 
-            $grouped[$date][] = $transaction;
+            $grouped[$group][] = $transaction;
         }
 
         return $grouped;
     }
 
-    public function getCurrentBalance(User $user)
+    public function getCurrentBalance()
     {
+        $user = Auth::user();
+
         return $user->transactions->sum('amount');
     }
 
@@ -44,7 +57,7 @@ class Transactions
     {
         $transaction = new Transaction();
 
-        $transaction->user_id = $user->id;
+        $transaction->user_id = Auth::user()->id;
         $transaction->label = $data['label'];
         $transaction->amount = $data['amount'];
         $transaction->occurred_at = Carbon::createFromTimeString($data['occurred_at'])->format('Y-m-d H:i:s');
