+++ public/3.js
@@ -1,18 +1,14 @@
 (window["webpackJsonp"] = window["webpackJsonp"] || []).push([[3],{
 
-/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/dashboard/new-entry.vue?vue&type=script&lang=js&":
-/*!******************************************************************************************************************************************************************************!*\
-  !*** ./node_modules/babel-loader/lib??ref--4-0!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/dashboard/new-entry.vue?vue&type=script&lang=js& ***!
-  \******************************************************************************************************************************************************************************/
+/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/dashboard/edit-entry.vue?vue&type=script&lang=js&":
+/*!*******************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/babel-loader/lib??ref--4-0!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/dashboard/edit-entry.vue?vue&type=script&lang=js& ***!
+  \*******************************************************************************************************************************************************************************/
 /*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
 __webpack_require__.r(__webpack_exports__);
-/* harmony import */ var vue_datetime__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! vue-datetime */ "./node_modules/vue-datetime/dist/vue-datetime.js");
-/* harmony import */ var vue_datetime__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(vue_datetime__WEBPACK_IMPORTED_MODULE_0__);
-/* harmony import */ var vue_datetime_dist_vue_datetime_css__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! vue-datetime/dist/vue-datetime.css */ "./node_modules/vue-datetime/dist/vue-datetime.css");
-/* harmony import */ var vue_datetime_dist_vue_datetime_css__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(vue_datetime_dist_vue_datetime_css__WEBPACK_IMPORTED_MODULE_1__);
 //
 //
 //
@@ -42,34 +38,36 @@ __webpack_require__.r(__webpack_exports__);
 //
 //
 //
-
-
+//
 /* harmony default export */ __webpack_exports__["default"] = ({
-  name: "new-entry",
+  name: "edit-entry",
   props: {
-    action: {
+    transaction: {
       required: true,
-      type: String
+      type: Object
     }
   },
-  components: {
-    Datetime: vue_datetime__WEBPACK_IMPORTED_MODULE_0__["Datetime"]
-  },
-  data: function data() {
-    return {
-      showModal: false,
-      heading: 'Add Balance Entry',
-      occurredAt: ''
-    };
+  methods: {
+    updateTransaction: function updateTransaction() {
+      var _this = this;
+
+      axios.patch('/transaction/' + this.transaction.id, {
+        'label': this.transaction.label,
+        'amount': this.transaction.amount,
+        'occurred_at': this.transaction.occurred_at
+      }).then(function (response) {
+        _this.$emit('edited');
+      });
+    }
   }
 });
 
 /***/ }),
 
-/***/ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/dashboard/new-entry.vue?vue&type=template&id=6e5ee54f&":
-/*!**********************************************************************************************************************************************************************************************************************!*\
-  !*** ./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/dashboard/new-entry.vue?vue&type=template&id=6e5ee54f& ***!
-  \**********************************************************************************************************************************************************************************************************************/
+/***/ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/dashboard/edit-entry.vue?vue&type=template&id=abff95d6&scoped=true&":
+/*!***********************************************************************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/dashboard/edit-entry.vue?vue&type=template&id=abff95d6&scoped=true& ***!
+  \***********************************************************************************************************************************************************************************************************************************/
 /*! exports provided: render, staticRenderFns */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
@@ -81,126 +79,157 @@ var render = function() {
   var _vm = this
   var _h = _vm.$createElement
   var _c = _vm._self._c || _h
-  return _c(
-    "span",
-    [
+  return _c("div", { staticClass: "px-4 py-2 mt-5" }, [
+    _c("div", { staticClass: "grid grid-cols-9 gap-6 mb-5" }, [
+      _c("div", { staticClass: "col-span-6 sm:col-span-4" }, [
+        _c(
+          "label",
+          {
+            staticClass: "block text-sm font-medium text-gray-700 uppercase",
+            attrs: { for: "label" }
+          },
+          [_vm._v("Label")]
+        ),
+        _vm._v(" "),
+        _c("input", {
+          directives: [
+            {
+              name: "model",
+              rawName: "v-model",
+              value: _vm.transaction.label,
+              expression: "transaction.label"
+            }
+          ],
+          staticClass:
+            "mt-1 text-black focus:ring-indigo-500 focus:border-indigo-500 block w-full shadow-sm sm:text-sm border border-gray-300 rounded-md",
+          attrs: {
+            type: "text",
+            name: "label",
+            id: "label",
+            autocomplete: "label",
+            required: "required"
+          },
+          domProps: { value: _vm.transaction.label },
+          on: {
+            input: function($event) {
+              if ($event.target.composing) {
+                return
+              }
+              _vm.$set(_vm.transaction, "label", $event.target.value)
+            }
+          }
+        })
+      ]),
+      _vm._v(" "),
       _c(
-        "a",
+        "div",
+        { staticClass: "col-span-6 sm:col-span-3" },
+        [
+          _c(
+            "label",
+            {
+              staticClass: "block text-sm font-medium text-gray-700 uppercase",
+              attrs: { for: "occurred_at" }
+            },
+            [_vm._v("Date")]
+          ),
+          _vm._v(" "),
+          _c("datetime", {
+            attrs: {
+              type: "datetime",
+              "hidden-name": "occurred_at",
+              required: true,
+              "use12-hour": true,
+              "input-class":
+                "mt-1 text-black focus:ring-indigo-500 focus:border-indigo-500 block w-full shadow-sm sm:text-sm border border-gray-300 rounded-md"
+            },
+            model: {
+              value: _vm.transaction.occurred_at,
+              callback: function($$v) {
+                _vm.$set(_vm.transaction, "occurred_at", $$v)
+              },
+              expression: "transaction.occurred_at"
+            }
+          })
+        ],
+        1
+      ),
+      _vm._v(" "),
+      _c("div", { staticClass: "col-span-6 sm:col-span-2" }, [
+        _c(
+          "label",
+          {
+            staticClass: "block text-sm font-medium text-gray-700 uppercase",
+            attrs: { for: "amount" }
+          },
+          [_vm._v("Amount")]
+        ),
+        _vm._v(" "),
+        _c("input", {
+          directives: [
+            {
+              name: "model",
+              rawName: "v-model",
+              value: _vm.transaction.amount,
+              expression: "transaction.amount"
+            }
+          ],
+          staticClass:
+            "mt-1 text-black focus:ring-indigo-500 focus:border-indigo-500 block w-full shadow-sm sm:text-sm border border-gray-300 rounded-md",
+          attrs: {
+            type: "text",
+            name: "amount",
+            id: "amount",
+            autocomplete: "amount",
+            required: "required"
+          },
+          domProps: { value: _vm.transaction.amount },
+          on: {
+            input: function($event) {
+              if ($event.target.composing) {
+                return
+              }
+              _vm.$set(_vm.transaction, "amount", $event.target.value)
+            }
+          }
+        })
+      ])
+    ]),
+    _vm._v(" "),
+    _c("div", { staticClass: "py-3 sm:flex sm:flex-row-reverse" }, [
+      _c(
+        "button",
         {
           staticClass:
-            "flex items-center mr-4 px-3 py-2 bg-blue-700 rounded-md text-white text-xs font-bold uppercase tracking-tight",
-          attrs: { href: "#" },
+            "w-full inline-flex justify-center rounded-md border border-transparent shadow-sm px-4 py-2 bg-blue-600 text-base font-medium text-white hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 sm:ml-3 sm:w-auto sm:text-sm",
+          attrs: { type: "submit" },
           on: {
             click: function($event) {
-              _vm.showModal = true
+              $event.preventDefault()
+              return _vm.updateTransaction($event)
             }
           }
         },
-        [_vm._v("\n        Add Entry\n    ")]
+        [_vm._v("\n            Update Entry\n        ")]
       ),
       _vm._v(" "),
       _c(
-        "modal",
+        "button",
         {
-          attrs: {
-            showing: _vm.showModal,
-            heading: _vm.heading,
-            action: _vm.action
+          staticClass:
+            "mt-3 w-full inline-flex justify-center rounded-md border border-gray-300 shadow-sm px-4 py-2 bg-white text-base font-medium text-gray-700 hover:bg-gray-50 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 sm:mt-0 sm:ml-3 sm:w-auto sm:text-sm",
+          attrs: { type: "button" },
+          on: {
+            click: function($event) {
+              $event.preventDefault()
+              return _vm.$emit("cancel")
+            }
           }
         },
-        [
-          _c("div", { staticClass: "grid grid-cols-9 gap-6 mb-5" }, [
-            _c("div", { staticClass: "col-span-6 sm:col-span-4" }, [
-              _c(
-                "label",
-                {
-                  staticClass:
-                    "block text-sm font-medium text-gray-700 uppercase",
-                  attrs: { for: "label" }
-                },
-                [_vm._v("Label")]
-              ),
-              _vm._v(" "),
-              _c("input", {
-                staticClass:
-                  "mt-1 text-black focus:ring-indigo-500 focus:border-indigo-500 block w-full shadow-sm sm:text-sm border border-gray-300 rounded-md",
-                attrs: {
-                  type: "text",
-                  name: "label",
-                  id: "label",
-                  autocomplete: "label",
-                  required: "required"
-                }
-              })
-            ]),
-            _vm._v(" "),
-            _c(
-              "div",
-              { staticClass: "col-span-6 sm:col-span-3" },
-              [
-                _c(
-                  "label",
-                  {
-                    staticClass:
-                      "block text-sm font-medium text-gray-700 uppercase",
-                    attrs: { for: "occurred_at" }
-                  },
-                  [_vm._v("Date")]
-                ),
-                _vm._v(" "),
-                _c("datetime", {
-                  attrs: {
-                    type: "datetime",
-                    "hidden-name": "occurred_at",
-                    required: true,
-                    "use12-hour": true,
-                    "input-class":
-                      "mt-1 text-black focus:ring-indigo-500 focus:border-indigo-500 block w-full shadow-sm sm:text-sm border border-gray-300 rounded-md"
-                  },
-                  model: {
-                    value: _vm.occurredAt,
-                    callback: function($$v) {
-                      _vm.occurredAt = $$v
-                    },
-                    expression: "occurredAt"
-                  }
-                })
-              ],
-              1
-            ),
-            _vm._v(" "),
-            _c("div", { staticClass: "col-span-6 sm:col-span-2" }, [
-              _c(
-                "label",
-                {
-                  staticClass:
-                    "block text-sm font-medium text-gray-700 uppercase",
-                  attrs: { for: "amount" }
-                },
-                [_vm._v("Amount")]
-              ),
-              _vm._v(" "),
-              _c("input", {
-                staticClass:
-                  "mt-1 text-black focus:ring-indigo-500 focus:border-indigo-500 block w-full shadow-sm sm:text-sm border border-gray-300 rounded-md",
-                attrs: {
-                  type: "text",
-                  name: "amount",
-                  id: "amount",
-                  autocomplete: "amount",
-                  required: "required"
-                }
-              })
-            ])
-          ]),
-          _vm._v(" "),
-          _vm._t("default")
-        ],
-        2
+        [_vm._v("\n            Cancel\n        ")]
       )
-    ],
-    1
-  )
+    ])
+  ])
 }
 var staticRenderFns = []
 render._withStripped = true
@@ -209,17 +238,17 @@ render._withStripped = true
 
 /***/ }),
 
-/***/ "./resources/js/components/dashboard/new-entry.vue":
-/*!*********************************************************!*\
-  !*** ./resources/js/components/dashboard/new-entry.vue ***!
-  \*********************************************************/
+/***/ "./resources/js/components/dashboard/edit-entry.vue":
+/*!**********************************************************!*\
+  !*** ./resources/js/components/dashboard/edit-entry.vue ***!
+  \**********************************************************/
 /*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
 __webpack_require__.r(__webpack_exports__);
-/* harmony import */ var _new_entry_vue_vue_type_template_id_6e5ee54f___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./new-entry.vue?vue&type=template&id=6e5ee54f& */ "./resources/js/components/dashboard/new-entry.vue?vue&type=template&id=6e5ee54f&");
-/* harmony import */ var _new_entry_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./new-entry.vue?vue&type=script&lang=js& */ "./resources/js/components/dashboard/new-entry.vue?vue&type=script&lang=js&");
+/* harmony import */ var _edit_entry_vue_vue_type_template_id_abff95d6_scoped_true___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./edit-entry.vue?vue&type=template&id=abff95d6&scoped=true& */ "./resources/js/components/dashboard/edit-entry.vue?vue&type=template&id=abff95d6&scoped=true&");
+/* harmony import */ var _edit_entry_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./edit-entry.vue?vue&type=script&lang=js& */ "./resources/js/components/dashboard/edit-entry.vue?vue&type=script&lang=js&");
 /* empty/unused harmony star reexport *//* harmony import */ var _node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../../../node_modules/vue-loader/lib/runtime/componentNormalizer.js */ "./node_modules/vue-loader/lib/runtime/componentNormalizer.js");
 
 
@@ -229,50 +258,50 @@ __webpack_require__.r(__webpack_exports__);
 /* normalize component */
 
 var component = Object(_node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__["default"])(
-  _new_entry_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__["default"],
-  _new_entry_vue_vue_type_template_id_6e5ee54f___WEBPACK_IMPORTED_MODULE_0__["render"],
-  _new_entry_vue_vue_type_template_id_6e5ee54f___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"],
+  _edit_entry_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__["default"],
+  _edit_entry_vue_vue_type_template_id_abff95d6_scoped_true___WEBPACK_IMPORTED_MODULE_0__["render"],
+  _edit_entry_vue_vue_type_template_id_abff95d6_scoped_true___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"],
   false,
   null,
-  null,
+  "abff95d6",
   null
   
 )
 
 /* hot reload */
 if (false) { var api; }
-component.options.__file = "resources/js/components/dashboard/new-entry.vue"
+component.options.__file = "resources/js/components/dashboard/edit-entry.vue"
 /* harmony default export */ __webpack_exports__["default"] = (component.exports);
 
 /***/ }),
 
-/***/ "./resources/js/components/dashboard/new-entry.vue?vue&type=script&lang=js&":
-/*!**********************************************************************************!*\
-  !*** ./resources/js/components/dashboard/new-entry.vue?vue&type=script&lang=js& ***!
-  \**********************************************************************************/
+/***/ "./resources/js/components/dashboard/edit-entry.vue?vue&type=script&lang=js&":
+/*!***********************************************************************************!*\
+  !*** ./resources/js/components/dashboard/edit-entry.vue?vue&type=script&lang=js& ***!
+  \***********************************************************************************/
 /*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
 __webpack_require__.r(__webpack_exports__);
-/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_new_entry_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/babel-loader/lib??ref--4-0!../../../../node_modules/vue-loader/lib??vue-loader-options!./new-entry.vue?vue&type=script&lang=js& */ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/dashboard/new-entry.vue?vue&type=script&lang=js&");
-/* empty/unused harmony star reexport */ /* harmony default export */ __webpack_exports__["default"] = (_node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_new_entry_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__["default"]); 
+/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_edit_entry_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/babel-loader/lib??ref--4-0!../../../../node_modules/vue-loader/lib??vue-loader-options!./edit-entry.vue?vue&type=script&lang=js& */ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/dashboard/edit-entry.vue?vue&type=script&lang=js&");
+/* empty/unused harmony star reexport */ /* harmony default export */ __webpack_exports__["default"] = (_node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_edit_entry_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__["default"]); 
 
 /***/ }),
 
-/***/ "./resources/js/components/dashboard/new-entry.vue?vue&type=template&id=6e5ee54f&":
-/*!****************************************************************************************!*\
-  !*** ./resources/js/components/dashboard/new-entry.vue?vue&type=template&id=6e5ee54f& ***!
-  \****************************************************************************************/
+/***/ "./resources/js/components/dashboard/edit-entry.vue?vue&type=template&id=abff95d6&scoped=true&":
+/*!*****************************************************************************************************!*\
+  !*** ./resources/js/components/dashboard/edit-entry.vue?vue&type=template&id=abff95d6&scoped=true& ***!
+  \*****************************************************************************************************/
 /*! exports provided: render, staticRenderFns */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
 __webpack_require__.r(__webpack_exports__);
-/* harmony import */ var _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_new_entry_vue_vue_type_template_id_6e5ee54f___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!../../../../node_modules/vue-loader/lib??vue-loader-options!./new-entry.vue?vue&type=template&id=6e5ee54f& */ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/dashboard/new-entry.vue?vue&type=template&id=6e5ee54f&");
-/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "render", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_new_entry_vue_vue_type_template_id_6e5ee54f___WEBPACK_IMPORTED_MODULE_0__["render"]; });
+/* harmony import */ var _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_edit_entry_vue_vue_type_template_id_abff95d6_scoped_true___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!../../../../node_modules/vue-loader/lib??vue-loader-options!./edit-entry.vue?vue&type=template&id=abff95d6&scoped=true& */ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/dashboard/edit-entry.vue?vue&type=template&id=abff95d6&scoped=true&");
+/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "render", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_edit_entry_vue_vue_type_template_id_abff95d6_scoped_true___WEBPACK_IMPORTED_MODULE_0__["render"]; });
 
-/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "staticRenderFns", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_new_entry_vue_vue_type_template_id_6e5ee54f___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"]; });
+/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "staticRenderFns", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_edit_entry_vue_vue_type_template_id_abff95d6_scoped_true___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"]; });
