+++ public/4.js
@@ -1,14 +1,18 @@
 (window["webpackJsonp"] = window["webpackJsonp"] || []).push([[4],{
 
-/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/dashboard/transaction-group.vue?vue&type=script&lang=js&":
-/*!**************************************************************************************************************************************************************************************!*\
-  !*** ./node_modules/babel-loader/lib??ref--4-0!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/dashboard/transaction-group.vue?vue&type=script&lang=js& ***!
-  \**************************************************************************************************************************************************************************************/
+/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/dashboard/new-entry.vue?vue&type=script&lang=js&":
+/*!******************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/babel-loader/lib??ref--4-0!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/dashboard/new-entry.vue?vue&type=script&lang=js& ***!
+  \******************************************************************************************************************************************************************************/
 /*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
 __webpack_require__.r(__webpack_exports__);
+/* harmony import */ var vue_datetime__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! vue-datetime */ "./node_modules/vue-datetime/dist/vue-datetime.js");
+/* harmony import */ var vue_datetime__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(vue_datetime__WEBPACK_IMPORTED_MODULE_0__);
+/* harmony import */ var vue_datetime_dist_vue_datetime_css__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! vue-datetime/dist/vue-datetime.css */ "./node_modules/vue-datetime/dist/vue-datetime.css");
+/* harmony import */ var vue_datetime_dist_vue_datetime_css__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(vue_datetime_dist_vue_datetime_css__WEBPACK_IMPORTED_MODULE_1__);
 //
 //
 //
@@ -22,71 +26,59 @@ __webpack_require__.r(__webpack_exports__);
 //
 //
 //
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+
+
 /* harmony default export */ __webpack_exports__["default"] = ({
-  name: "transaction-group",
+  name: "new-entry",
   props: {
-    date: {
+    action: {
       required: true,
       type: String
     },
-    data: {
+    processing: {
       required: true,
-      type: Array
+      type: String
     }
+  },
+  components: {
+    Datetime: vue_datetime__WEBPACK_IMPORTED_MODULE_0__["Datetime"]
   },
   data: function data() {
     return {
-      transactions: this.data
+      showModal: false,
+      heading: 'Add Balance Entry',
+      occurredAt: ''
     };
   },
-  watch: {
-    data: function data() {
-      this.transactions = this.data;
-    }
-  },
-  methods: {
-    removeTransaction: function removeTransaction(id) {
-      var _this = this;
-
-      this.transactions.forEach(function (transaction) {
-        if (transaction.id === id) {
-          _this.transactions.splice(_this.transactions.indexOf(transaction), 1);
-
-          _this.$emit('remove');
-        }
-      });
-    },
-    updateTransaction: function updateTransaction() {
-      this.$emit('update');
-    }
-  },
   computed: {
-    total: function total() {
-      var total = 0;
-      this.transactions.forEach(function (transaction) {
-        total += parseFloat(transaction.amount);
-      });
-      return total;
-    },
-    formattedAmount: function formattedAmount() {
-      return Math.abs(this.total).toFixed(0);
-    },
-    fraction: function fraction() {
-      // I am sad to do this, but don't wanna waste time here :[
-      return this.total.toFixed(2).toString().split('.')[1];
-    },
-    amountClass: function amountClass() {
-      return this.total > 0 ? 'text-green-500' : 'text-gray-500';
+    isDisabled: function isDisabled() {
+      return this.processing === 'true';
     }
   }
 });
 
 /***/ }),
 
-/***/ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/dashboard/transaction-group.vue?vue&type=template&id=0daee69a&":
-/*!******************************************************************************************************************************************************************************************************************************!*\
-  !*** ./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/dashboard/transaction-group.vue?vue&type=template&id=0daee69a& ***!
-  \******************************************************************************************************************************************************************************************************************************/
+/***/ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/dashboard/new-entry.vue?vue&type=template&id=6e5ee54f&":
+/*!**********************************************************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/dashboard/new-entry.vue?vue&type=template&id=6e5ee54f& ***!
+  \**********************************************************************************************************************************************************************************************************************/
 /*! exports provided: render, staticRenderFns */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
@@ -98,56 +90,126 @@ var render = function() {
   var _vm = this
   var _h = _vm.$createElement
   var _c = _vm._self._c || _h
-  return _vm.transactions.length
-    ? _c(
-        "div",
-        { staticClass: "mb-8" },
+  return _c(
+    "span",
+    [
+      _c(
+        "button",
+        {
+          staticClass:
+            "flex items-center mr-4 px-3 py-2 bg-blue-700 disabled:opacity-50 rounded-md text-white text-xs font-bold uppercase tracking-tight",
+          attrs: { type: "button", disabled: _vm.isDisabled },
+          on: {
+            click: function($event) {
+              _vm.showModal = true
+            }
+          }
+        },
+        [_vm._v("\n        Add Entry\n    ")]
+      ),
+      _vm._v(" "),
+      _c(
+        "modal",
+        {
+          attrs: {
+            showing: _vm.showModal,
+            heading: _vm.heading,
+            action: _vm.action
+          }
+        },
         [
-          _c("div", { staticClass: "flex items-center mb-4" }, [
-            _c(
-              "span",
-              {
+          _c("div", { staticClass: "grid grid-cols-9 gap-6 mb-5" }, [
+            _c("div", { staticClass: "col-span-6 sm:col-span-4" }, [
+              _c(
+                "label",
+                {
+                  staticClass:
+                    "block text-sm font-medium text-gray-700 uppercase",
+                  attrs: { for: "label" }
+                },
+                [_vm._v("Label")]
+              ),
+              _vm._v(" "),
+              _c("input", {
                 staticClass:
-                  "flex-grow text-gray-500 font-bold text-sm uppercase tracking-tight"
-              },
-              [_vm._v(_vm._s(_vm.date))]
-            ),
+                  "mt-1 text-black focus:ring-indigo-500 focus:border-indigo-500 block w-full shadow-sm sm:text-sm border border-gray-300 rounded-md",
+                attrs: {
+                  type: "text",
+                  name: "label",
+                  id: "label",
+                  autocomplete: "label",
+                  required: "required"
+                }
+              })
+            ]),
             _vm._v(" "),
             _c(
-              "span",
-              {
-                staticClass: "text-lg text-gray-500 font-bold",
-                class: _vm.amountClass
-              },
+              "div",
+              { staticClass: "col-span-6 sm:col-span-3" },
               [
-                _vm._v(
-                  "\n            " +
-                    _vm._s(this.total > 0 ? "+ " : "- ") +
-                    "\n            $" +
-                    _vm._s(_vm.formattedAmount) +
-                    "."
+                _c(
+                  "label",
+                  {
+                    staticClass:
+                      "block text-sm font-medium text-gray-700 uppercase",
+                    attrs: { for: "occurred_at" }
+                  },
+                  [_vm._v("Date")]
                 ),
-                _c("span", { staticClass: "text-sm" }, [
-                  _vm._v(_vm._s(_vm.fraction))
-                ])
-              ]
-            )
+                _vm._v(" "),
+                _c("datetime", {
+                  attrs: {
+                    type: "datetime",
+                    "hidden-name": "occurred_at",
+                    required: true,
+                    "use12-hour": true,
+                    "input-class":
+                      "mt-1 text-black focus:ring-indigo-500 focus:border-indigo-500 block w-full shadow-sm sm:text-sm border border-gray-300 rounded-md"
+                  },
+                  model: {
+                    value: _vm.occurredAt,
+                    callback: function($$v) {
+                      _vm.occurredAt = $$v
+                    },
+                    expression: "occurredAt"
+                  }
+                })
+              ],
+              1
+            ),
+            _vm._v(" "),
+            _c("div", { staticClass: "col-span-6 sm:col-span-2" }, [
+              _c(
+                "label",
+                {
+                  staticClass:
+                    "block text-sm font-medium text-gray-700 uppercase",
+                  attrs: { for: "amount" }
+                },
+                [_vm._v("Amount")]
+              ),
+              _vm._v(" "),
+              _c("input", {
+                staticClass:
+                  "mt-1 text-black focus:ring-indigo-500 focus:border-indigo-500 block w-full shadow-sm sm:text-sm border border-gray-300 rounded-md",
+                attrs: {
+                  type: "text",
+                  name: "amount",
+                  id: "amount",
+                  autocomplete: "amount",
+                  required: "required"
+                }
+              })
+            ])
           ]),
           _vm._v(" "),
-          _vm._l(_vm.transactions, function(transaction) {
-            return _c("transaction", {
-              key: transaction.id,
-              attrs: { transaction: transaction },
-              on: {
-                remove: _vm.removeTransaction,
-                update: _vm.updateTransaction
-              }
-            })
-          })
+          _vm._t("default")
         ],
         2
       )
-    : _vm._e()
+    ],
+    1
+  )
 }
 var staticRenderFns = []
 render._withStripped = true
@@ -156,17 +218,17 @@ render._withStripped = true
 
 /***/ }),
 
-/***/ "./resources/js/components/dashboard/transaction-group.vue":
-/*!*****************************************************************!*\
-  !*** ./resources/js/components/dashboard/transaction-group.vue ***!
-  \*****************************************************************/
+/***/ "./resources/js/components/dashboard/new-entry.vue":
+/*!*********************************************************!*\
+  !*** ./resources/js/components/dashboard/new-entry.vue ***!
+  \*********************************************************/
 /*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
 __webpack_require__.r(__webpack_exports__);
-/* harmony import */ var _transaction_group_vue_vue_type_template_id_0daee69a___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./transaction-group.vue?vue&type=template&id=0daee69a& */ "./resources/js/components/dashboard/transaction-group.vue?vue&type=template&id=0daee69a&");
-/* harmony import */ var _transaction_group_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./transaction-group.vue?vue&type=script&lang=js& */ "./resources/js/components/dashboard/transaction-group.vue?vue&type=script&lang=js&");
+/* harmony import */ var _new_entry_vue_vue_type_template_id_6e5ee54f___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./new-entry.vue?vue&type=template&id=6e5ee54f& */ "./resources/js/components/dashboard/new-entry.vue?vue&type=template&id=6e5ee54f&");
+/* harmony import */ var _new_entry_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./new-entry.vue?vue&type=script&lang=js& */ "./resources/js/components/dashboard/new-entry.vue?vue&type=script&lang=js&");
 /* empty/unused harmony star reexport *//* harmony import */ var _node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../../../node_modules/vue-loader/lib/runtime/componentNormalizer.js */ "./node_modules/vue-loader/lib/runtime/componentNormalizer.js");
 
 
@@ -176,9 +238,9 @@ __webpack_require__.r(__webpack_exports__);
 /* normalize component */
 
 var component = Object(_node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__["default"])(
-  _transaction_group_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__["default"],
-  _transaction_group_vue_vue_type_template_id_0daee69a___WEBPACK_IMPORTED_MODULE_0__["render"],
-  _transaction_group_vue_vue_type_template_id_0daee69a___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"],
+  _new_entry_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__["default"],
+  _new_entry_vue_vue_type_template_id_6e5ee54f___WEBPACK_IMPORTED_MODULE_0__["render"],
+  _new_entry_vue_vue_type_template_id_6e5ee54f___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"],
   false,
   null,
   null,
@@ -188,38 +250,38 @@ var component = Object(_node_modules_vue_loader_lib_runtime_componentNormalizer_
 
 /* hot reload */
 if (false) { var api; }
-component.options.__file = "resources/js/components/dashboard/transaction-group.vue"
+component.options.__file = "resources/js/components/dashboard/new-entry.vue"
 /* harmony default export */ __webpack_exports__["default"] = (component.exports);
 
 /***/ }),
 
-/***/ "./resources/js/components/dashboard/transaction-group.vue?vue&type=script&lang=js&":
-/*!******************************************************************************************!*\
-  !*** ./resources/js/components/dashboard/transaction-group.vue?vue&type=script&lang=js& ***!
-  \******************************************************************************************/
+/***/ "./resources/js/components/dashboard/new-entry.vue?vue&type=script&lang=js&":
+/*!**********************************************************************************!*\
+  !*** ./resources/js/components/dashboard/new-entry.vue?vue&type=script&lang=js& ***!
+  \**********************************************************************************/
 /*! exports provided: default */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
 __webpack_require__.r(__webpack_exports__);
-/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_transaction_group_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/babel-loader/lib??ref--4-0!../../../../node_modules/vue-loader/lib??vue-loader-options!./transaction-group.vue?vue&type=script&lang=js& */ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/dashboard/transaction-group.vue?vue&type=script&lang=js&");
-/* empty/unused harmony star reexport */ /* harmony default export */ __webpack_exports__["default"] = (_node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_transaction_group_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__["default"]); 
+/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_new_entry_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/babel-loader/lib??ref--4-0!../../../../node_modules/vue-loader/lib??vue-loader-options!./new-entry.vue?vue&type=script&lang=js& */ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/dashboard/new-entry.vue?vue&type=script&lang=js&");
+/* empty/unused harmony star reexport */ /* harmony default export */ __webpack_exports__["default"] = (_node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_new_entry_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__["default"]); 
 
 /***/ }),
 
-/***/ "./resources/js/components/dashboard/transaction-group.vue?vue&type=template&id=0daee69a&":
-/*!************************************************************************************************!*\
-  !*** ./resources/js/components/dashboard/transaction-group.vue?vue&type=template&id=0daee69a& ***!
-  \************************************************************************************************/
+/***/ "./resources/js/components/dashboard/new-entry.vue?vue&type=template&id=6e5ee54f&":
+/*!****************************************************************************************!*\
+  !*** ./resources/js/components/dashboard/new-entry.vue?vue&type=template&id=6e5ee54f& ***!
+  \****************************************************************************************/
 /*! exports provided: render, staticRenderFns */
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
 __webpack_require__.r(__webpack_exports__);
-/* harmony import */ var _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_transaction_group_vue_vue_type_template_id_0daee69a___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!../../../../node_modules/vue-loader/lib??vue-loader-options!./transaction-group.vue?vue&type=template&id=0daee69a& */ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/dashboard/transaction-group.vue?vue&type=template&id=0daee69a&");
-/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "render", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_transaction_group_vue_vue_type_template_id_0daee69a___WEBPACK_IMPORTED_MODULE_0__["render"]; });
+/* harmony import */ var _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_new_entry_vue_vue_type_template_id_6e5ee54f___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!../../../../node_modules/vue-loader/lib??vue-loader-options!./new-entry.vue?vue&type=template&id=6e5ee54f& */ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/dashboard/new-entry.vue?vue&type=template&id=6e5ee54f&");
+/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "render", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_new_entry_vue_vue_type_template_id_6e5ee54f___WEBPACK_IMPORTED_MODULE_0__["render"]; });
 
-/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "staticRenderFns", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_transaction_group_vue_vue_type_template_id_0daee69a___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"]; });
+/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "staticRenderFns", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_new_entry_vue_vue_type_template_id_6e5ee54f___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"]; });
