+++ resources/js/app.js
@@ -4,6 +4,7 @@ Vue.component('transaction', () => import('./components/dashboard/transaction'))
 Vue.component('transaction-group', () => import('./components/dashboard/transaction-group'));
 Vue.component('new-entry', () => import('./components/dashboard/new-entry'));
 Vue.component('edit-entry', () => import('./components/dashboard/edit-entry'));
+Vue.component('bulk-import', () => import('./components/dashboard/bulk-import'));
 Vue.component('balance', () => import('./components/dashboard/balance'));
 Vue.component('modal', require('./components/Modal.vue').default);
