+++ resources/js/components/modal.vue
@@ -1,13 +1,13 @@
 <template>
     <Transition name="fade">
-        <div v-if="showing" class="fixed z-10 inset-0 overflow-y-auto">
+        <div v-if="open" class="fixed z-10 inset-0 overflow-y-auto">
             <div class="flex items-end justify-center min-h-screen pt-4 px-4 pb-20 text-center sm:block sm:p-0">
                 <div class="fixed inset-0 transition-opacity" aria-hidden="true">
                     <div class="absolute inset-0 bg-gray-500 opacity-75"></div>
                 </div>
                 <span class="hidden sm:inline-block sm:align-middle sm:h-screen" aria-hidden="true">&#8203;</span>
                 <div class="inline-block align-bottom bg-white rounded-lg text-left shadow-xl transform transition-all sm:my-8 sm:align-middle sm:max-w-2xl sm:w-full" role="dialog" aria-modal="true" aria-labelledby="modal-headline">
-                    <form :action="action" method="POST">
+                    <form :action="action" method="POST" :enctype="upload ? 'multipart/form-data' : 'application/x-www-form-urlencoded'">
                         <h2 class="xl:font-extrabold text-black text-xl px-4 pt-5 pb-4 sm:p-6 sm:pb-4">
                             {{ heading }}
                         </h2>
@@ -47,13 +47,26 @@ export default {
         showing: {
             required: true,
             type: Boolean
+        },
+        upload: {
+            required: false,
+            type: Boolean
         }
     },
-
+    data() {
+        return {
+            open: this.showing
+        }
+    },
+    watch: {
+        showing() {
+            this.open = this.showing
+        }
+    },
     methods: {
         close() {
             this.$emit('close');
-            this.showing = false;
+            this.open = false;
         },
     }
 };
